backend-api/requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==8.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-jwt-extended==4.5.3
marshmallow==3.20.1
